Add Cardano/HFC Byron mainnet config
Add configs for running a node that connects to Byron mainnet but
talks the Cardano/HFC protocol, so that it delivers Cardano blocks
over the local domain socket. This configuration requires the
existing Byron mainnet genesis but also a dummy Shelley genesis
(it needs to exist and parse correctly, but is otherwise null).